refactored into seperate components
