+++ src/feedback/FeedbackForm.js
@@ -1,13 +1,10 @@
 import Box from '@mui/material/Box';
 import { styled } from '@mui/system';
-import Stack from '@mui/material/Stack';
 import React, { useState, useReducer } from 'react';
-import Button from '@mui/material/Button';
 import DialogSubmit from './DialogSubmit';
 import TopDecor from './top_decoration.png';
-import TextField from '@mui/material/TextField';
-import Typography from '@mui/material/Typography';
-import LoadingButton from '@mui/lab/LoadingButton';
+import Inputs from './Inputs';
+
 const Image = styled('img')({
   height: '100px',
   width: '100%',
@@ -39,14 +36,6 @@ function FeedbackForm() {
   //  }, initialState);
 
   // const { status, error, data } = state;
-
-  const handleChange = (event, key) => {
-    const config = {
-      section: (el) => setSection(el),
-      feedback: (el) => setFeedback(el),
-    };
-    config[event.target.id](event.target.value);
-  };
 
   const handleSubmit = (event) => {
     event.preventDefault();
@@ -98,76 +87,14 @@ function FeedbackForm() {
         sx={{
           p: 1,
         }}>
-        <Stack
-          spacing={2}
-          justifyContent='center'
-          alignItems='center'
-          sx={{ ml: 4, mr: 4, mb: 4 }}>
-          <Typography
-            component='h1'
-            sx={{ fontWeight: 700, fontSize: '30px', color: '#575F6C' }}>
-            Feedback
-          </Typography>
-          <Typography
-            component='h1'
-            align='left'
-            sx={{ fontWeight: 700, fontSize: '18px', color: '#3F4957' }}>
-            We're always looking for ways to make VIZE better!
-          </Typography>
-          <Typography
-            component='h1'
-            align='left'
-            sx={{ fontWeight: 500, fontSize: '18px', color: '#3F4957' }}>
-            What Section of VIZE do you have a suggestion for?
-          </Typography>
-          <TextField
-            fullWidth
-            defaultValue='Small'
-            variant='filled'
-            id='section'
-            label='Section'
-            value={section}
-            required
-            onChange={(e) => handleChange(e)}
-            InputProps={{ disableUnderline: true }}
-          />
-          <Typography
-            align='left'
-            component='p'
-            sx={{ fontWeight: 500, fontSize: '18px' }}>
-            What Can we do better?
-          </Typography>
-          <TextField
-            fullWidth
-            multiline
-            rows={8}
-            defaultValue='feedback'
-            variant='filled'
-            id='feedback'
-            label='Feedback'
-            required
-            value={feedback}
-            onChange={(e) => handleChange(e)}
-            InputProps={{ disableUnderline: true }}
-          />
-          <Stack direction={{ xs: 'column', sm: 'row' }} spacing={2}>
-            <Button
-              variant='outlined'
-              size='large'
-              sx={{ borderRadius: '20px', width: '150px' }}>
-              Back
-            </Button>
-            <LoadingButton
-              sx={{ borderRadius: '20px', fontWeight: 600, width: '150px' }}
-              type='submit'
-              size='large'
-              loading={loading}
-              loadingPosition='start'
-              variant='contained'>
-              Submit
-            </LoadingButton>
-          </Stack>
-        </Stack>
+        <Inputs
+          section={section}
+          setSection={setSection}
+          feedback={feedback}
+          setFeedback={setFeedback}
+          loading={loading}
+          setLoading={setLoading}
+        />
       </Box>
       <DialogSubmit open={open} setOpen={setOpen} />
     </Box>
